requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# DAG Builder dependencies
jinja2>=3.1.0
pyyaml>=6.0
jsonschema>=4.0.0

# Additional ML/Data dependencies
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
